requirements.txt
fastapi==0.115.13
pydantic==2.11.7
requests==2.32.4
torch==2.7.1 --index-url https://download.pytorch.org/whl/cpu
transformers==4.52.4
numpy==2.3.0
sqlalchemy==2.0.41
streamlit==1.46.0
chromadb==1.0.13
llama-cpp-python==0.3.9
uvicorn==0.34.3
pywhispercpp~=1.3.2
streamlit-audiorecorder~=0.0.6
scikit-learn==1.7.0
ffmpeg~=1.4
python-multipart~=0.0.20
websocket-client~=1.9.0
pymongo~=4.14.0
python-dotenv~=1.1.1